pyproject.toml
[tool.poetry]
name = "lipidetective"
version = "0.1.0"
description = "A deep learning framework for the identification of molecular lipid species from tandem mass spectra."
authors = ["Vivian Wuerf", "Nikolai Koehler", "Florian Molnar", "Lisa Hahnefeld", "Robert Gurke", "Michael Witting", "Josch K. Pauling"]
license = "BSD 3-Clause License"
readme = "README.md"
packages = [{ include = "lipidetective", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.0"
pytorch-lightning = "^2.1.2"
ray = {version = "2.6.3", extras = ["tune"]}
numpy = "^1.26.3"
matplotlib = "^3.8.0"
h5py = "^3.9.0"
pandas = "^2.1.4"
seaborn = "^0.13.2"
sphinx = "^7.2.6"
sphinx-togglebutton = "^0.3.2"
sphinx_rtd_theme = "^2.0.0"
torchmetrics = "^1.2.1"
umap-learn = "^0.5.5"
pyyaml = "^6.0.1"
scikit-learn = "^1.3.1"
wandb = "0.16.3"
pyteomics = "4.6.3"
lxml = "5.1.0"
holoviews = "^1.18.1"

[tool.poetry.group.dev.dependencies]
poetry-core = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
lipidetective = "lipidetective.lipidetective:main"
